requirements.txt
torchdyn==1.0.6
timm==0.9.2
tensorboard==2.12.2
einops==0.4.1
matplotlib==3.5.3
mujoco==2.2.1
numpy==1.23.1
opencv_python==4.7.0.72
pandas==1.4.3
Pillow==9.5.0
scipy==1.9.0
tqdm==4.64.0